requirements.txt
openai
requests
ipykernel
python-dotenv
pandas
snowflake-sqlalchemy
marvin